+++ pkg/log/http-in-response.go
@@ -6,10 +6,6 @@ import (
 	"time"
 )
 
-var hlcInResponse *headerObfuscator = newHeaderObfuscator().
-	obfuscateHeaders([]string{"Authorization", "WWW-Authenticate"}).
-	ignoreHeaders([]string{"X-Real-Ip", "Content-Encoding", "Content-Type", "Accept-Encoding", "Content-Length", "Date"})
-
 type inResponseLog struct {
 	Timestamp       time.Time           `json:"timestamp"`
 	LogLevel        logLevel            `json:"log-level"`